+++ src/main/java/fr/estecka/variantscit/modules/PotionTypeModule.java
@@ -1,13 +1,13 @@
-package fr.estecka.variantscit.modules.potion_effect;
+package fr.estecka.variantscit.modules;
 
-import fr.estecka.variantscit.api.IVariantProvider;
+import fr.estecka.variantscit.api.ISimpleCitModule;
 import net.minecraft.component.DataComponentTypes;
 import net.minecraft.component.type.PotionContentsComponent;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
 
 public class PotionTypeModule
-implements IVariantProvider
+implements ISimpleCitModule
 {
 	@Override
 	public Identifier GetItemVariant(ItemStack stack){
